changed floats to decimals

+++ btrade/accounts/models.py
@@ -5,7 +5,7 @@ from django.db.models.signals import post_save
 # Create your models here.
 class UserProfile(models.Model):
     user = models.OneToOneField(User)
-    currency = models.FloatField()
+    currency = models.DecimalField(decimal_places=2, max_digits=15)
     earned_currency = models.IntegerField(default=0)
 
     def __str__(self):
